+++ src/components/Cart/Cart.js
@@ -2,6 +2,7 @@ import { useContext } from "react";
 import CartContext from "../../store/cart-context";
 import Overlay from "../UI/Overlay";
 import styles from "./Cart.module.css";
+import CartItem from "./CartItem";
 
 const Cart = (props) => {
 	const ctx = useContext(CartContext);
@@ -9,11 +10,7 @@ const Cart = (props) => {
 	let items = [];
 	for (const key in ctx.items) {
 		const element = ctx.items[key];
-		const item = (
-			<li id={element.id} key={element.id} className={styles.item}>
-				{element.name}
-			</li>
-		);
+		const item = <CartItem key={element.id} item={element} />;
 		items = [...items, item];
 	}